Refactor host names
It is beneficial for the host names of the nodes to
match the DNS names one would assign them. This makes
it easier to use web UIs, since some of them
(Mesos, Marathon...) use the servers' host names to
build URLs in the UI. It is also easier to manage.

1) Added .short_name.domain to every host name. This
paves the way for configuring DNS records based on
the host names.

2) Removed number formatting; the number formatting
limited the number of worker nodes to 999, and was
one-indexed. Now, we use ${count.index} raw, so that
the list is zero-indexed and there's no limit on
the number of worker nodes one can have.

3) Remove short_name as a prefix, instead using it
as a subdomain. This lets the DNS structure be
organised from least-significant to most-significant,
where the name of the node is less significant than
the short_name of the data center, which in turn is
less significant than the top level domain name.

This changes the host names from this scheme:

mi-control-01
mi-control-02
mi-control-03
mi-worker-001
mi-worker-002
...
mi-worker-015

To this scheme:

control-0.dc0.example.com
control-1.dc0.example.com
control-2.dc0.example.com
worker-0.dc0.example.com
worker-1.dc0.example.com
...
worker-14.dc0.example.com